Refactor index.d.ts: split it in multiple files (in the types directory)
