+++ package.json
@@ -7,10 +7,10 @@
   "exports": {
     ".": {
       "import": "./src/index.js",
-      "types": "./index.d.ts"
+      "types": "./types/index.d.ts"
     }
   },
-  "types": "index.d.ts",
+  "types": "types/index.d.ts",
   "keywords": [
     "pushpad",
     "web-push",
