+++ types/Sender.d.ts
@@ -1,0 +1,7 @@
+export interface Sender {
+  id?: number;
+  name?: string;
+  vapid_private_key?: string;
+  vapid_public_key?: string;
+  created_at?: string;
+}
